Refactoring: extracting some checks from LogMessageArchive to CertificateValidator and CertificateHelper
